requirements.txt
absl-py==2.1.0
ansi2html==1.9.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clifford==1.4.0
colorama==0.4.6
comm==0.2.0
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.47.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
install==1.3.5
ipython==8.18.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_core==5.6.1
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
mpl-toolkits.clifford==0.0.3
nbformat==5.7.0
nest-asyncio==1.5.8
numba==0.58.1
numpy==1.26.2
open3d==0.17.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==3.20.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pywin32==306
referencing==0.32.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.16.2
scipy==1.11.4
six==1.16.0
sounddevice==0.4.6
sparse==0.14.0
stack-data==0.6.3
tenacity==8.2.3
tk==0.1.0
traitlets==5.14.0
trimesh==4.0.8
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
widgetsnbextension==4.0.9
zipp==3.17.0
